% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplets_generation.R
\name{generate_all_multiplets}
\alias{generate_all_multiplets}
\title{Generate All Candidate Variable Multiplets}
\usage{
generate_all_multiplets(n_vars, k = 3)
}
\arguments{
\item{n_vars}{Total number of variables.}

\item{k}{Order of candidate multiplets (\eqn{k \ge 3}).}
}
\value{
A list of integer vectors, each representing a variable
subset of size \code{k}.
}
\description{
Generates all possible variable subsets of a given order k (k ≥ 3).
}
\details{
This function implements the brute-force candidate generation
strategy described by Marinazzo et al. (2024), where every
possible combination of variables of size \code{k} is evaluated.
Although exhaustive, the number of candidate multiplets grows
combinatorially with the number of variables \eqn{\binom{p}{k}}

where \eqn{p} is the number of variables.

For large psychometric datasets this approach may become
computationally expensive, motivating more selective
candidate generation procedures such as those proposed by
Possenti et al. (2026).
}
\examples{
generate_all_multiplets(
  n_vars = 5,
  k = 3
)

}
\references{
Marinazzo, D., Van Roozendaal, J., Rosas, F. E., Stella, M.,
Comolatti, R., Colenbier, N., Stramaglia, S., & Rosseel, Y. (2024).
An information-theoretic approach to build hypergraphs in psychometrics.
\emph{Behavior Research Methods}, \emph{56}(7), 8057-8079.
https://doi.org/10.3758/s13428-024-02471-8
}
