% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_step_testing.R
\name{permutation_Oinfo}
\alias{permutation_Oinfo}
\title{Permutation Test for O-information}
\usage{
permutation_Oinfo(data, vars, n_perm = 1000, cor_method = "pearson")
}
\arguments{
\item{data}{Normalized data matrix or data frame.}

\item{vars}{Integer vector indicating variables belonging
to the candidate multiplet.}

\item{n_perm}{Number of permutations.}

\item{cor_method}{Correlation method passed to
\code{compute_correlation()}.}
}
\value{
A list containing:
\describe{
\item{obs_O}{Observed O-information.}
\item{pval}{Permutation p-value.}
}
}
\description{
Performs a permutation-based significance test for the
O-information of a candidate variable subset.
}
\details{
This procedure corresponds to the first stage
of the hypergraph construction framework proposed by
Possenti et al. (2026).

The null hypothesis assumes that the observed higher-order
dependency arises from chance. A null distribution is
generated by independently permuting each variable,
thereby preserving univariate distributions while
destroying multivariate dependence structure.

The empirical p-value is computed as:

\deqn{
p =
\frac{
\#(|\Omega_{perm}| \ge |\Omega_{obs}|)
}{
N_{perm}
}
}

where \eqn{\Omega_{obs}} is the observed O-information
and \eqn{\Omega_{perm}} are values obtained from
permuted datasets.
}
\references{
Possenti, F., Girelli, L., Tieri, P., & Petti, M. (2026). Multiplex
Hypergraph Modeling of Higher Order Structures in Psychometric Networks
(Version 1). arXiv. https://doi.org/10.48550/ARXIV.2604.22744
}
