% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_hypergraph.R
\name{print.psychHypergraph}
\alias{print.psychHypergraph}
\title{Print method for psychHypergraph objects}
\usage{
\method{print}{psychHypergraph}(x, ...)
}
\arguments{
\item{x}{A psychHypergraph object}

\item{...}{Additional arguments}
}
\description{
Print method for psychHypergraph objects
}
