% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{print.psychHypergraphClustering}
\alias{print.psychHypergraphClustering}
\title{Print method for psychHypergraphClustering}
\usage{
\method{print}{psychHypergraphClustering}(x, ...)
}
\arguments{
\item{x}{psychHypergraphClustering object}

\item{...}{Additional arguments}
}
\description{
Print method for psychHypergraphClustering
}
