% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{suggest_k}
\alias{suggest_k}
\title{Suggest the Number of Clusters Using the Eigengap Heuristic}
\usage{
suggest_k(psych_hg, max_K = 10, weighted_l = FALSE, ...)
}
\arguments{
\item{psych_hg}{A \code{psychHypergraph} object.}

\item{max_K}{Maximum number of clusters to evaluate.}

\item{weighted_l}{Logical; whether to use a weighted
incidence matrix.}

\item{...}{Additional arguments.}
}
\value{
A list containing:

\describe{
\item{suggested_K}{Recommended number of clusters.}
\item{eigenvalues}{Evaluated eigenvalues.}
\item{eigengaps}{Consecutive eigenvalue differences.}
\item{eigenvalues_all}{Full eigenvalue spectrum.}
}
}
\description{
Estimate an appropriate number of clusters for hypergraph
spectral clustering using the eigengap heuristic.
}
\details{
The method computes Laplacian eigenvalues and identifies
the largest gap between consecutive eigenvalues.

A large eigengap indicates a natural partition of the
hypergraph into communities.

Let
\eqn{\lambda_1 \le \lambda_2 \le \cdots}
denote Laplacian eigenvalues.

The eigengap for cluster number \eqn{k} is:

\deqn{
gap_k = \lambda_{k+1} - \lambda_k
}

The suggested number of clusters corresponds to the
largest observed eigengap.
}
\examples{
\dontrun{

suggest_k(hg)

k_info <- suggest_k(
  hg,
  max_K = 8
)

k_info$suggested_K

}

}
\seealso{
\code{\link{spectral_clustering}}
}
