% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_hypergraph.R
\name{summary.psychHypergraph}
\alias{summary.psychHypergraph}
\title{Summary method for psychHypergraph objects}
\usage{
\method{summary}{psychHypergraph}(object, ...)
}
\arguments{
\item{object}{A psychHypergraph object}

\item{...}{Additional arguments}
}
\description{
Summary method for psychHypergraph objects
}
